src/backend/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

aiosmtplib==4.0.2
app==0.0.1
bcrypt==5.0.0
fastapi==0.117.1
httpx==0.28.1
icalendar==6.3.1
itsdangerous==2.2.0
Jinja2==3.1.6
pydantic==2.11.9
email-validator==2.2.0
pytest==8.4.2
python-dateutil==2.9.0.post0
pytz==2025.2
slowapi==0.1.9
SQLAlchemy==2.0.43
starlette==0.48.0
structlog==25.4.0
uvicorn==0.36.0
weasyprint==66.0
asyncpg==0.30.0
greenlet==3.2.4
# WARNING(pigar): some manual fixes might be required as pigar has detected duplicate requirements for the same import name (possibly for different submodules).
# WARNING(pigar): the following duplicate requirements are for the import name: jose

python-jose==3.5.0
# WARNING(pigar): the following duplicate requirements are for the import name: passlib
passlib==1.7.4
passlib-fork==1.7.4.post2
# WARNING(pigar): the following duplicate requirements are for the import name: psutil
psutil==5.9.8
# WARNING(pigar): the following duplicate requirements are for the import name: pydantic_settings
pydantic-settings==2.11.0
